57 cleared
Insert a closed-interval into a list of closed intervals.
In the beginning the list is non-overlapping and sorted in increasing
order.
Insert and merge the lists s.t. result is non-overlapping and sorted and
in increasing order, plus that the union of intervals in result is
exactly the same set as that of the new interval modulo the original
list.

Iterators are kinda fun :)