Number of prime factors alone won't do. Eratosthenes it is.
The gist is that using the H combination.

similar to nCr and nPr, nHr is a shorthand for the size of some specific
combinations, in particular its the number of non-negative integer
solutions to the problem:

$x_1 + x_2 + \cdots + x_n = r > 0$

viewing the plus signs as indistinguishable placeholders, this problem
is then effectively asking mCs where $m = (n-1) + r$ and $s = r$, i.e.
nHr = (n+r-1)Cr

With the annotation above, we see that for each prime factor $p$ with
multiplicity $r$ and the array being length $n$, corresponding number of
ways is just nHr.
Multiply them all and we have the answer.

Hence we need exactly all the multiplicity of the prime factors, rather
than the total number of prime factors (including multiplicity).