[Palindrome] Manacher's algorithm
Probably O(n) since the `try_for_each` part
Beating 100% in time and 50% in space

Sometimes it's better to just give up and check the comments, since
it's impossible for any individual to re-invent all the great algorithms
in the history.
Manacher's algorithm is such a case; though I've indeed had a glimpse
into how it's invented.
Anyhow for the general case the insertion of special characters for even
palindromes is quite genious.

I suppose there are few people trying Rust on Leetcode...