[binary search] sometimes elegance comes from asking the right question
In particular, to check for which sub-slice is the culprit, we compare
its first and last element.

The gist is to note that if the slice is the culprit iff its first
element is larger than the last element.

We're searching for the minimum, which implies that the first element
may also be discarded!

Which makes asking if the latter sub-slice being the culprit a tad bit
more efficient than asking the former sub-slice, since to take advantage
of this trick, it would require one more assignment.

Micro optimizations are cool!

The symmetry breakage comes from we're asking for the index of the minimum
in a strictly ascending array (but rotated).

https://leetcode.com/problems/find-minimum-in-rotated-sorted-array/solutions/158940/beat-100-very-simple-python-very-detailed-explanation