[binary search] worst case linear proof, and just find min in that case
Sometimes the eldest trick in the book has more to offer than you
might've imagined.

Remember that you might have explained that why naive quick sort is
actually square time, when sorting an array into non-decreasing the
algorithm is given an input that's non-increasing?

It's called adversary argument, and in this case it's easy to apply:
for any N and any deterministic algorithm,
there exists array in which all elements are 1 except a 0 that makes the
algorithm run in linear time.

Construct the array as following:
being a deterministic algorithm,
when fed with an array with length N,
it tries to poke some (distinct) indices one by one.

Utilize this basic fact, we construct the array by for every access,
fill in that entry with 1 and return that 1,
except for the last access where we put the sneaky 0.

In the algorithm's PoV, all the elements it had encountered say they are
1, so it has no better way but to try to access other element, only to
find they are again 1, till it had exhausted the array and met that 0.

It's just like the fact you can't possibly find min without having
looked at all the elements: that one guy might just be the actual min.